Fix an oobread when pressing the backspace in the r2 shell (#16844)
* Regression inrtoduced when moving from int to size_t in r_str_utf8_charsize_prev

str.c:2043:8: runtime error: addition of unsigned offset to 0x00010c7de67f overflowed to 0x00010c7de67e
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior str.c:2043:8 in

+++ libr/cons/dietline.c
@@ -862,7 +862,7 @@ static inline void __delete_next_char() {
 static inline void __delete_prev_char() {
 	if (I.buffer.index < I.buffer.length) {
 		if (I.buffer.index > 0) {
-			int len = r_str_utf8_charsize_prev (I.buffer.data + I.buffer.index, I.buffer.index);
+			size_t len = r_str_utf8_charsize_prev (I.buffer.data + I.buffer.index, I.buffer.index);
 			I.buffer.index -= len;
 			memmove (I.buffer.data + I.buffer.index,
 				I.buffer.data + I.buffer.index + len,
